Address code-review findings on the RNG draws (#133)
Uniform(low, high) — three correctness defects in the range transform:

- P(positive side) was posW/(posW+negW), whose sum overflows to +inf once the
  interval is wide enough (|bound| > ~1.7e38), collapsing pPos to 0 and sending
  EVERY draw negative, silently. It is now computed scaled by max(posW, negW),
  so both terms are <= 1 and the denominator lands in [1,2] — no overflow at any
  width. Uniform(-2e38, 2e38) goes from 100% negative to 49.7% positive.
- The .Max(1e-30f) denominator guard pinned the denominator for narrow
  zero-straddling intervals and destroyed the split ratio: P(positive) measured
  0.0003 at +-1e-33 and 0.0000 at +-1e-35. The scaled form needs no guard at all
  (m is zero only when low = high = 0, which is not interior and lands in the
  discarded arm), so it is gone. +-1e-33 now measures 0.4967.
- The non-straddling case anchored at `low` unconditionally, so Uniform(low, 0)
  was low*(1-g) — reaching zero only as g->1, where 1-g is quantised at 2^-24.
  Every value was a multiple of 2^-24, i.e. the very cancellation the transform
  exists to avoid, mirrored. It now anchors at whichever endpoint is nearest
  zero and expands away from it.

VectorRange — the earlier fix was incomplete. The uint16 and uint32 arms carry
the identical narrowing bug (Cast<int16>/Cast<int32> wraps a large limit
negative, so Range yields an empty vector, silently, with no throw):
VectorRange(0u, 3000000000u, 1000000000u) returned empty. Both now range in a
width that holds the limit. The coverage exercised only [0,4), which cannot see
this; it now exercises each unsigned arm past its width's signed limit, and was
verified to fail when either fixed arm is reverted.

RngTestOracle.DrawNormal had zero call sites — the normal transform had no
host-oracle parity test at all, only self-generated goldens and moment ranges,
which is exactly the drift this oracle exists to prevent. It is now asserted in
TestInGraphBitsAndUniformDrawsMatchTheHostGeneratorBitExactly with a 1e-5
tolerance (Ln/Sqrt/Cos/Sin are EP-approximate, unlike the integer bits path and
the exactly-constructed uniform, which stay bit-exact).

NNLinearMatchesPyTorch's reference was regenerated from Shorokoo's own output by
this PR's bulk golden pass, silently downgrading the repo's only external
cross-check to a tautology while its comment still claimed PyTorch provenance.
Regenerated properly through tests/pytorch-reference/linear.py; PyTorch
independently reproduces the same values, so the check is external truth again.

Docs corrected where they described the pre-PR behaviour: the class summary
still described the 24-bit bit->float and the 1-u1 radius, and GeometricUniform
claimed to be exact without noting that while every produced VALUE is exact, the
DISTRIBUTION is a uniform truncated at 2^-41 with a doubled bottom octave (an
all-zero exponent field shares a bucket with a field of 1) and an open interval
at both ends.

Not addressed here: Uniform still evaluates both the magnitude and side-selector
draws on every call. Where is an elementwise select, not a branch — both
operands are always evaluated, so no EP or graph optimiser can elide the unused
one. Removing it needs build-time specialisation on constant bounds rather than
a runtime Where; tracked separately.

+++ src/Shorokoo/Global.Constructors.cs
@@ -331,9 +331,9 @@ namespace Shorokoo
         /// <summary>Creates a vector of values from start (inclusive) to limit (exclusive), stepping by delta.</summary>
         public static Vector<uint8> VectorRange(byte start, byte limit, byte delta) => OnnxOp.Range(Scalar(start).Cast<int32>(), Scalar(limit).Cast<int32>(), Scalar(delta).Cast<int32>()).Cast<uint8>().Vec();
         /// <summary>Creates a vector of values from start (inclusive) to limit (exclusive), stepping by delta.</summary>
-        public static Vector<uint16> VectorRange(ushort start, ushort limit, ushort delta) => OnnxOp.Range(Scalar(start).Cast<int16>(), Scalar(limit).Cast<int16>(), Scalar(delta).Cast<int16>()).Cast<uint16>().Vec();
+        public static Vector<uint16> VectorRange(ushort start, ushort limit, ushort delta) => OnnxOp.Range(Scalar(start).Cast<int32>(), Scalar(limit).Cast<int32>(), Scalar(delta).Cast<int32>()).Cast<uint16>().Vec();
         /// <summary>Creates a vector of values from start (inclusive) to limit (exclusive), stepping by delta.</summary>
-        public static Vector<uint32> VectorRange(uint start, uint limit, uint delta) => OnnxOp.Range(Scalar(start).Cast<int32>(), Scalar(limit).Cast<int32>(), Scalar(delta).Cast<int32>()).Cast<uint32>().Vec();
+        public static Vector<uint32> VectorRange(uint start, uint limit, uint delta) => OnnxOp.Range(Scalar(start).Cast<int64>(), Scalar(limit).Cast<int64>(), Scalar(delta).Cast<int64>()).Cast<uint32>().Vec();
         /// <summary>Creates a vector of values from start (inclusive) to limit (exclusive), stepping by delta.</summary>
         public static Vector<uint64> VectorRange(ulong start, ulong limit, ulong delta) => OnnxOp.Range(Scalar(start).Cast<int64>(), Scalar(limit).Cast<int64>(), Scalar(delta).Cast<int64>()).Cast<uint64>().Vec();
